LPS-128599 Merge JDKLoggerTestUtil and Log4JLoggerTestUtil to LoggerTestUtil. Let configureJDKLogger() return LogCapture, instead of JDKLogCapture.
