+++ portal-test/src/com/liferay/portal/kernel/test/CaptureHandler.java
@@ -28,7 +28,7 @@ import java.util.logging.Logger;
 /**
  * @author Shuyang Zhou
  * @deprecated As of Cavanaugh (7.4.x), replaced by {@link
- *             JDKLoggerTestUtil.CaptureHandler}
+ *             com.liferay.portal.test.log.LoggerTestUtil.JDKLogCapture}
  */
 @Deprecated
 public class CaptureHandler extends Handler implements Closeable {
