+++ portal-test/src/com/liferay/portal/kernel/test/JDKLoggerTestUtil.java
@@ -25,7 +25,7 @@ import java.util.logging.Logger;
 /**
  * @author Shuyang Zhou
  * @deprecated As of Cavanaugh (7.4.x), replaced by {@link
- *             com.liferay.portal.test.log.jdk.JDKLoggerTestUtil}
+ *             com.liferay.portal.test.log.LoggerTestUtil}
  */
 @Deprecated
 public class JDKLoggerTestUtil {
