+++ portal-test/src/com/liferay/portal/test/log/LogEvent.java
@@ -22,7 +22,7 @@ import org.apache.log4j.spi.ThrowableInformation;
 /**
  * @author Tina Tian
  * @deprecated As of Cavanaugh (7.4.x), replaced by {@link
- *              Log4JLoggerTestUtil.LogEvent}
+ *              LoggerTestUtil.Log4JLogEntry}
  */
 @Deprecated
 public class LogEvent {
